refact qpe to harmonize qiskit and pennylane
